+++ docs/changelog.md
@@ -9,6 +9,7 @@
 ## Версия 1.5.0
 - Учитывать всех исполнителей трека в фильтр-функциях `remove*`, `match*`, `dedup*`.
 - [getArtistsTracks](/func?id=getartiststracks) и [getArtistsAlbums](/func?id=getartistsalbums) получили параметр `isFlat`, позволяющий сгруппировать результат по исполнителю. По аналогии с [getArtistsTopTracks](/func?id=getartiststoptracks). Поведение по умолчанию не изменилось, код менять не нужно.
+- `rangeTracks` может фильтровать по типу альбома.
 - Все get-запросы из `SpotifyRequest` содержат параметр локали ([подробнее](https://github.com/Chimildic/goofy/discussions/79#discussioncomment-1019029)).
 
 ## Версия 1.4.9