+++ docs/changelog.md
@@ -10,6 +10,7 @@
 - Учитывать всех исполнителей трека в фильтр-функциях `remove*`, `match*`, `dedup*`.
 - [getArtistsTracks](/func?id=getartiststracks) и [getArtistsAlbums](/func?id=getartistsalbums) получили параметр `isFlat`, позволяющий сгруппировать результат по исполнителю. По аналогии с [getArtistsTopTracks](/func?id=getartiststoptracks). Поведение по умолчанию не изменилось, код менять не нужно.
 - `rangeTracks` может фильтровать по типу альбома.
+- `dedupTracks` получил новый аргумент для контроля отклонения продолжительности у треков с идентичными названиями. Подробнее [здесь](https://github.com/Chimildic/goofy/discussions/116).
 - Все get-запросы из `SpotifyRequest` содержат параметр локали ([подробнее](https://github.com/Chimildic/goofy/discussions/79#discussioncomment-1019029)).
 
 ## Версия 1.4.9